Code refactoring for Dynamic Deck Dynamo
