+++ bundle/pom.xml
@@ -86,15 +86,16 @@
 							*
 						</Import-Package>
 						<Sling-Model-Packages>
-							com.adobe.acs.commons.cloudconfig,
-							com.adobe.acs.commons.marketo,
-							com.adobe.acs.commons.redirectmaps.models,
-							com.adobe.acs.commons.version.model,
-							com.adobe.acs.commons.wcm.comparisons.model,
-							com.adobe.acs.commons.workflow.bulk.execution.model,
-							com.adobe.acs.commons.mcp.model,
-							com.adobe.acs.commons.reports.models
-						</Sling-Model-Packages>
+                            com.adobe.acs.commons.cloudconfig,
+                            com.adobe.acs.commons.marketo,
+                            com.adobe.acs.commons.redirectmaps.models,
+                            com.adobe.acs.commons.version.model,
+                            com.adobe.acs.commons.wcm.comparisons.model,
+                            com.adobe.acs.commons.workflow.bulk.execution.model,
+                            com.adobe.acs.commons.mcp.model,
+                            com.adobe.acs.commons.reports.models,
+                            com.adobe.acs.commons.indesign.dynamicdeckdynamo.models
+                        </Sling-Model-Packages>
 					</instructions>
 				</configuration>
 			</plugin>
