+++ bundle/src/main/java/com/adobe/acs/commons/indesign/dynamicdeckdynamo/constants/DeckDynamoIDSConstants.java
@@ -1,4 +1,4 @@
-package com.adobe.acs.commons.indesign.deckdynamo.constants;
+package com.adobe.acs.commons.indesign.dynamicdeckdynamo.constants;
 
 /**
  * This class will be used to store Indesign Server Specific Constants
