+++ bundle/src/main/java/com/adobe/acs/commons/indesign/dynamicdeckdynamo/models/DeckDynamoInitiatorPageModel.java
@@ -1,4 +1,4 @@
-package com.adobe.acs.commons.indesign.deckdynamo.models;
+package com.adobe.acs.commons.indesign.dynamicdeckdynamo.models;
 
 import com.adobe.acs.commons.mcp.form.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
