+++ bundle/src/main/java/com/adobe/acs/commons/indesign/dynamicdeckdynamo/osgiconfigurations/DeckDynamoConfiguration.java
@@ -1,4 +1,4 @@
-package com.adobe.acs.commons.indesign.deckdynamo.osgiconfigurations;
+package com.adobe.acs.commons.indesign.dynamicdeckdynamo.osgiconfigurations;
 
 import org.osgi.service.metatype.annotations.AttributeDefinition;
 import org.osgi.service.metatype.annotations.AttributeType;
@@ -8,13 +8,10 @@ import org.osgi.service.metatype.annotations.ObjectClassDefinition;
  * These are some important configurations for generation InDesign deck dynamically.
  */
 @ObjectClassDefinition(
-        name = "Deck Dynamo Configurations",
+        name = "Dynamic Deck Dynamo Configurations",
         description = "These are some important configurations for generation InDesign deck dynamically."
 )
 public @interface DeckDynamoConfiguration {
-
-    @AttributeDefinition(name = "Template Root Path", description = "This path is the root path of InDesign templates", type = AttributeType.STRING)
-    public String templateRootPath() default "/content/dam/dynamic-deck-dynamo/templates";
 
     @AttributeDefinition(name = "Placeholder Image Path", description = "This placeholder image is used if deck image is not available", type = AttributeType.STRING)
     public String placeholderImagePath() default "/content/dam/dynamic-deck-dynamo/placeholder-images/APPAREL_PLACEHOLDER_TRANSPARENT.png";
