+++ bundle/src/main/java/com/adobe/acs/commons/indesign/dynamicdeckdynamo/services/XMLGeneratorService.java
@@ -1,7 +1,7 @@
-package com.adobe.acs.commons.indesign.deckdynamo.services;
+package com.adobe.acs.commons.indesign.dynamicdeckdynamo.services;
 
-import com.adobe.acs.commons.indesign.deckdynamo.exception.DeckDynamoException;
-import com.adobe.acs.commons.indesign.deckdynamo.pojos.XMLResourceIterator;
+import com.adobe.acs.commons.indesign.dynamicdeckdynamo.exception.DeckDynamoException;
+import com.adobe.acs.commons.indesign.dynamicdeckdynamo.pojos.XMLResourceIterator;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
